@@ -3,11 +3,14 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
+using System.Text.Json;
 using common.Extensions;
 using kernel.Abstraction;
+using kernel.Extension;
+using kernel.Internal;
 using kernel.Service;
+using Microsoft.AspNetCore.Mvc.Authorization;
 using Microsoft.Extensions.DependencyInjection;
-using Newtonsoft.Json;
 
 namespace kernel
 {
@@ -17,7 +20,7 @@ namespace kernel
         public IServiceCollection Services { get; }
         public TypeFinder TypeFinder { get; }
 
-        public Kernel(Assembly applicationScope, IServiceCollection services = null)
+        public Kernel(Assembly applicationScope, IServiceCollection? services = null)
         {
             ApplicationScope = applicationScope;
             Services = services ?? new ServiceCollection();
@@ -28,8 +31,9 @@ namespace kernel
 
         public void Boot()
         {
+            ProcessInternalRecorders();
             ProcessRecorders();
-            ProcessGlobalFilters();
+            ProcessMvc();
 
             Services.AddHttpContextAccessor();
         }
@@ -38,32 +42,50 @@ namespace kernel
         {
             if (!File.Exists(envFilename))
             {
-                throw new FileNotFoundException($"Unable to read environment file. File with name {envFilename} was not found.");
+                throw new FileNotFoundException(
+                    $"Unable to read environment file. File with name {envFilename} was not found.");
             }
 
             File.ReadAllText(envFilename);
 
-            var json = JsonConvert.DeserializeObject<Dictionary<string, string>>(File.ReadAllText(envFilename));
+            var json = JsonSerializer.Deserialize<Dictionary<string, string>>(File.ReadAllText(envFilename));
 
-            foreach (var pair in json)
+            foreach (var (key, value) in json)
             {
-                Environment.SetEnvironmentVariable(pair.Key, pair.Value);
+                Environment.SetEnvironmentVariable(key, value);
             }
         }
 
         private void ProcessRecorders()
         {
             TypeFinder.FindTypes(t => t.IsSubclassOf(typeof(AbstractServiceRecorder)) && !t.IsAbstract)
-                .Select(t => Activator.CreateInstance(t) as AbstractServiceRecorder)
-                .Foreach(recorder => recorder.Process(Services));
+                .Foreach(recorder =>
+                {
+                    Services.AddTransient(recorder)
+                        .GetService<AbstractServiceRecorder>(recorder)
+                        .Process(Services);
+                });
         }
 
-        private void ProcessGlobalFilters()
+        private void ProcessMvc()
         {
-            var mvcBuilder = Services.AddMvc();
+            var mvcBuilder = Services.AddRouting().AddMvc();
 
-            TypeFinder.FindTypes(t => t.ImplementedInterfaces.Contains(typeof(IGlobalFilter)))
+            mvcBuilder.AddMvcOptions(options => options.Filters.Add(new AuthorizeFilter()));
+
+            TypeFinder.FindTypes(IsGlobalFilter)
                 .Foreach(filter => mvcBuilder.AddMvcOptions(o => o.Filters.Add(filter)));
+        }
+
+        private void ProcessInternalRecorders()
+        {
+            Services.AddTransient(typeof(ResolvingAttributeServiceRecorder));
+            Services.GetService<ResolvingAttributeServiceRecorder>()!.Process(Services);
+        }
+
+        private bool IsGlobalFilter(TypeInfo type)
+        {
+            return type.ImplementedInterfaces.Contains(typeof(IGlobalFilter));
         }
     }
 }
