@@ -9,20 +9,19 @@ namespace command_runner
         public Assembly CliScope { get; }
         public IServiceCollection Services { get; }
 
-
         public CommandRunner(Assembly cliScope, IServiceCollection services)
         {
             CliScope = cliScope;
             Services = services ?? new ServiceCollection();
         }
 
-        public void Execute(string defaultCommand = null)
+        public void Execute(string? defaultCommand = null)
         {
             var consoleManager = new ConsoleManager(CliScope, Services);
 
             if (0 != defaultCommand?.Length)
             {
-                consoleManager.InvokeWithoutStart(defaultCommand);
+                consoleManager.InvokeWithoutStart(defaultCommand!);
             }
             else
             {
