@@ -1,4 +1,14 @@
-$HEADER$namespace $NAMESPACE$
+using System;
+
+namespace kernel.Attribute
 {
-  public class $CLASS$ {$END$}
+    public abstract class AbstractServiceAttribute : System.Attribute
+    {
+        public Type? Interface { get; }
+
+        protected AbstractServiceAttribute(Type? @interface = null)
+        {
+            Interface = @interface;
+        }
+    }
 }
