@@ -1,7 +1,11 @@
+using System;
+
 namespace kernel.Attribute
 {
-    public class ScopedAttribute
+    public class ScopedAttribute : AbstractServiceAttribute
     {
-        
+        public ScopedAttribute(Type? @interface = null) : base(@interface)
+        {
+        }
     }
 }
