@@ -1,7 +1,11 @@
+using System;
+
 namespace kernel.Attribute
 {
-    public class SingletoneAttribute
+    public class SingletonAttribute : AbstractServiceAttribute
     {
-        
+        public SingletonAttribute(Type? @interface = null) : base(@interface)
+        {
+        }
     }
 }
