@@ -1,7 +1,11 @@
+using System;
+
 namespace kernel.Attribute
 {
-    public class TransientAttribute
+    public class TransientAttribute : AbstractServiceAttribute
     {
-        
+        public TransientAttribute(Type? @interface = null) : base(@interface)
+        {
+        }
     }
 }
