@@ -2,6 +2,6 @@ namespace app.Application.Abstraction
 {
     public interface IPasswordEncoder
     {
-        
+        string Encrypt(string passwordToEncode);
     }
 }
