@@ -4,7 +4,7 @@ namespace app.Application.CQS
 {
     public abstract class AbstractUserAware
     {
-        protected User CurrentUser { get; private set; }
+        protected User? CurrentUser { get; private set; }
 
         public void SetCurrentUser(User user)
         {
