@@ -1,4 +1,5 @@
 using app.Application.Abstraction;
+using app.Application.CQS.Token.Exception;
 using app.Application.CQS.Token.Input;
 using app.Application.CQS.Token.Output;
 using app.Domain.User;
@@ -9,25 +10,27 @@ namespace app.Application.CQS.Token.Query
     {
         public IUserRepository UserRepository { get; }
         public IJwtTokenService TokenService { get; }
+        public IPasswordEncoder PasswordEncoder { get; }
 
-        public ObtainTokenQuery(IUserRepository userRepository, IJwtTokenService tokenService)
+        public ObtainTokenQuery(
+            IUserRepository userRepository,
+            IJwtTokenService tokenService,
+            IPasswordEncoder passwordEncoder
+        )
         {
             UserRepository = userRepository;
             TokenService = tokenService;
+            PasswordEncoder = passwordEncoder;
         }
 
         public TokenOutput Execute(TokenInput input)
         {
-//            var user = UserRepository.FindUserWithLogin(input.Login);
+            var encryptedPassword = PasswordEncoder.Encrypt(input.Password);
+            var user = UserRepository.FindUserWithLoginAndPassword(input.Login, encryptedPassword);
 
-            var user = new User("foobar", "barfoo", "barbaz");
-
-            return new TokenOutput
-            {
-                UserId = user.Id,
-                Roles = user.Roles,
-                Token = TokenService.CreateToken(user),
-            };
+            return null != user
+                ? new TokenOutput(user.Id, TokenService.CreateToken(user))
+                : throw UserNotFound.InvalidCredentials();
         }
     }
 }
