@@ -9,16 +9,16 @@ namespace app.Infrastructure.Common
     public struct DbAccessor
     {
         // Хост базы данных
-        public static string Host => Environment.GetEnvironmentVariable("DB_HOST");
+        public static string Host => Environment.GetEnvironmentVariable("DB_HOST") ?? "";
         
         // Имя базы данных
-        public static string Database => Environment.GetEnvironmentVariable("DB_NAME");
+        public static string Database => Environment.GetEnvironmentVariable("DB_NAME") ?? "";
         
         // Пользовать
-        public static string User => Environment.GetEnvironmentVariable("DB_USER");
+        public static string User => Environment.GetEnvironmentVariable("DB_USER") ?? "";
         
         // Пароль
-        public static string Password => Environment.GetEnvironmentVariable("DB_PASSWORD");
+        public static string Password => Environment.GetEnvironmentVariable("DB_PASSWORD") ?? "";
 
         // Готовая строка для connecion к БД.
         public static string ConnectionString => $"Server={Host};Port=5432;Database={Database};User Id={User};Password={Password};";
